Address comment, refactor and use exist ThreadUtils
