--- core/server/common/src/main/java/alluxio/ReflectionUtils.java
@@ -1,113 +1,0 @@
-/*
- * The Alluxio Open Foundation licenses this work under the Apache License, version 2.0
- * (the "License"). You may not use this work except in compliance with the License, which is
- * available at www.apache.org/licenses/LICENSE-2.0
- *
- * This software is distributed on an "AS IS" basis, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied, as more fully set forth in the License.
- *
- * See the NOTICE file distributed with this work for information regarding copyright ownership.
- */
-
-package alluxio;
-
-import org.slf4j.Logger;
-
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
-import java.io.UnsupportedEncodingException;
-import java.lang.management.ManagementFactory;
-import java.lang.management.ThreadInfo;
-import java.lang.management.ThreadMXBean;
-import java.nio.charset.Charset;
-
-/**
- * General reflection utils.
- */
-public final class ReflectionUtils {
-  private static final ThreadMXBean THREAD_BEAN =
-      ManagementFactory.getThreadMXBean();
-
-  private static String getTaskName(long id, String name) {
-    if (name == null) {
-      return Long.toString(id);
-    }
-    return id + " (" + name + ")";
-  }
-
-  /**
-   * Print all of the thread's information and stack traces.
-   *
-   * @param stream the stream to
-   * @param title  a string title for the stack trace
-   */
-  public static synchronized void printThreadInfo(PrintStream stream,
-      String title) {
-    final int STACK_DEPTH = 20;
-    boolean contention = THREAD_BEAN.isThreadContentionMonitoringEnabled();
-    long[] threadIds = THREAD_BEAN.getAllThreadIds();
-    stream.println("Process Thread Dump: " + title);
-    stream.println(threadIds.length + " active threads");
-    for (long tid : threadIds) {
-      ThreadInfo info = THREAD_BEAN.getThreadInfo(tid, STACK_DEPTH);
-      if (info == null) {
-        stream.println("  Inactive");
-        continue;
-      }
-      stream.println("Thread "
-          + getTaskName(info.getThreadId(), info.getThreadName()) + ":");
-      Thread.State state = info.getThreadState();
-      stream.println("  State: " + state);
-      stream.println("  Blocked count: " + info.getBlockedCount());
-      stream.println("  Waited count: " + info.getWaitedCount());
-      if (contention) {
-        stream.println("  Blocked time: " + info.getBlockedTime());
-        stream.println("  Waited time: " + info.getWaitedTime());
-      }
-      if (state == Thread.State.WAITING) {
-        stream.println("  Waiting on " + info.getLockName());
-      } else if (state == Thread.State.BLOCKED) {
-        stream.println("  Blocked on " + info.getLockName());
-        stream.println("  Blocked by "
-            + getTaskName(info.getLockOwnerId(), info.getLockOwnerName()));
-      }
-      stream.println("  Stack:");
-      for (StackTraceElement frame : info.getStackTrace()) {
-        stream.println("    " + frame.toString());
-      }
-    }
-    stream.flush();
-  }
-
-  private static long sPreviousLogTime = 0;
-
-  /**
-   * Log the current thread stacks at INFO level.
-   * @param log the logger that logs the stack trace
-   * @param title a descriptive title for the call stacks
-   * @param minInterval the minimum time from the last
-   */
-  public static void logThreadInfo(Logger log,
-      String title,
-      long minInterval) {
-    boolean dumpStack = false;
-    if (log.isInfoEnabled()) {
-      synchronized (ReflectionUtils.class) {
-        long now = System.nanoTime() / 1000000;
-        if (now - sPreviousLogTime >= minInterval * 1000) {
-          sPreviousLogTime = now;
-          dumpStack = true;
-        }
-      }
-      if (dumpStack) {
-        try {
-          ByteArrayOutputStream buffer = new ByteArrayOutputStream();
-          printThreadInfo(new PrintStream(buffer, false, "UTF-8"), title);
-          log.info(buffer.toString(Charset.defaultCharset().name()));
-        } catch (UnsupportedEncodingException ignored) {
-          // swallow the nit exception to keep silent.
-        }
-      }
-    }
-  }
-}
